google 검색 기능, 찜 기능 코드 분리
